Add passkey type, u2f types and rand helpers
Define a `Passkey` type which contains the private key stored as a
CoseKey for generic key types to be available without generics. CoseKey
also encodes all of the keys necessary information internally such as
the algorithm and curve type.

The `Passkey` type also contains the other required and optional
metadata expected from a passkey for usage internally by an
authenticator for filtering passkeys as well as filling the responses.

All the fields are public so that the internals can be transformed to
and from your bespoke storage schema which we don't wish to prescribe.
This only determines the minimum information needed by the other
`passkey` crates for their correct operations.

Defines all the u2f types for supporting operations in that protocol. This
module also contains methods for "upgrading" a u2f request to a ctap2
request. Most of the required information is available, except for rpId.
For this we set the applicationIdHash as the rpId. This means that
credentials created through u2f cannot be used in a ctap2 setting, nor
the opposite. The resulting credential, while under the type `Passkey`
are not proper passkeys as they are not discoverable.

Defines a couple helper methods wrapping the `rand` crate for number
generation used in tests and in the other `passkey-*` crates in order to
have a single implementation re-usable elsewhere.

Add a crate feature `serialize_bytes_as_base64_string` that changes the
serialization of the `Bytes` type in from an array of numbers or the
byte string, to a base64url string.

Fix documentation and added more documentation.