Release passkey-authenticator
## Authenticator type
This type is designed to be very flexible in usage while keeping the
actual specification implementation sealed. This is why it has so many
generics with traits backing those fields. One for user verification,
one for storage, and eventually one for cryptography.

## CredentialStore
This is the trait most consumers of this crate will need to be careful
with, this provides an entry point for the authenticator to find and
store passkeys. Note that the item it returns during filtering is an
associated type. This allows you to use the trait in other cases where
you want to use the same filtering logic but retain your own schema. So
long as it can be transformed into a Passkey, thats all the
authenticator needs.

Consumers must be carefull to filter correctly on the provided list and
RpId.

## Ctap2Api
This will be the main entry point for interacting with the
`Authenticator`. Its the CTAP2.0 compliant implementation for the basic
operations such as `make_credential` and `get_assertion`. This trait is
sealed to only the `Authenticator` as we want to keep a single
implementation of this trait and provide customizability through the
other pluggable aspects of the `Authenticator`.