Release passkey-client
This crate provides an entry point that takes a credential manager
request and uses its internal authenticator attachment to provide a
response to the `register` (credentials.create) and `authenticate`
(credentials.get).

Currently the Client type takes a static in-memory authenticator, but in
the future will be able to use passkey-transports to communicate with a
remote authenticator through the Ctap2Api trait.